refactor: move parameters functions and class
